+++ core/bones/randomSliceBone.py
@@ -55,13 +55,13 @@ class randomSliceBone(baseBone):
 
 			:param name: The property-name this bone has in its Skeleton (not the description!)
 			:type name: str
-			:param skel: The :class:`server.skeleton.Skeleton` instance this bone is part of
-			:type skel: :class:`server.skeleton.Skeleton`
-			:param dbFilter: The current :class:`server.db.Query` instance the filters should be applied to
-			:type dbFilter: :class:`server.db.Query`
+			:param skel: The :class:`viur.core.skeleton.Skeleton` instance this bone is part of
+			:type skel: :class:`viur.core.skeleton.Skeleton`
+			:param dbFilter: The current :class:`viur.core.db.Query` instance the filters should be applied to
+			:type dbFilter: :class:`viur.core.db.Query`
 			:param rawFilter: The dictionary of filters the client wants to have applied
 			:type rawFilter: dict
-			:returns: The modified :class:`server.db.Query`
+			:returns: The modified :class:`viur.core.db.Query`
 		"""
 
 		def applyFilterHook(dbfilter, property, value):
@@ -117,7 +117,7 @@ class randomSliceBone(baseBone):
 
 	def calculateInternalMultiQueryLimit(self, query, targetAmount):
 		"""
-			Tells :class:`server.db.Query` How much entries should be fetched in each subquery.
+			Tells :class:`viur.core.db.Query` How much entries should be fetched in each subquery.
 
 			:param targetAmount: How many entries shall be returned from db.Query
 			:type targetAmount: int
@@ -133,11 +133,11 @@ class randomSliceBone(baseBone):
 			:param dbFilter: The db.Query calling this function
 			:type: dbFilter: server.db.Query
 			:param result: The list of results for each subquery we've run
-			:type result: list of list of :class:`server.db.Entity`
+			:type result: list of list of :class:`viur.core.db.Entity`
 			:param targetAmount: How many results should be returned from db.Query
 			:type targetAmount: int
 			:return: list of elements which should be returned from db.Query
-			:rtype: list of :class:`server.db.Entity`
+			:rtype: list of :class:`viur.core.db.Entity`
 		"""
 		# res is a list of iterators at this point, chain them together
 		res = chain(*[list(x) for x in result])